Implement a MetricsFactory, move TsdMetrics to own file.
